+++ test/reject.rb
@@ -2,20 +2,20 @@
 
 require_relative 'test_helper'
 
-# confirm that :reject and its :exclude alias reject splits at the specified (1-based)
+# confirm that :reject and its :except alias reject splits at the specified (1-based)
 # indices
 
 describe 'reject' do
   ss = StringSplitter.new(default_delimiter: ':')
 
   test 'reject a positive position' do
-    result = ss.split('foo:bar:baz:quux', exclude: 1)
+    result = ss.split('foo:bar:baz:quux', except: 1)
     assert { result == %w[foo:bar baz quux] }
 
     result = ss.split('foo:bar:baz:quux', reject: 1)
     assert { result == %w[foo:bar baz quux] }
 
-    result = ss.rsplit('foo:bar:baz:quux', exclude: 1)
+    result = ss.rsplit('foo:bar:baz:quux', except: 1)
     assert { result == %w[foo bar baz:quux] }
 
     result = ss.rsplit('foo:bar:baz:quux', reject: 1)
@@ -23,13 +23,13 @@ describe 'reject' do
   end
 
   test 'reject a negative position' do
-    result = ss.split('foo:bar:baz:quux', exclude: -1)
+    result = ss.split('foo:bar:baz:quux', except: -1)
     assert { result == %w[foo bar baz:quux] }
 
     result = ss.split('foo:bar:baz:quux', reject: -1)
     assert { result == %w[foo bar baz:quux] }
 
-    result = ss.rsplit('foo:bar:baz:quux', exclude: -1)
+    result = ss.rsplit('foo:bar:baz:quux', except: -1)
     assert { result == %w[foo:bar baz quux] }
 
     result = ss.rsplit('foo:bar:baz:quux', reject: -1)
@@ -37,13 +37,13 @@ describe 'reject' do
   end
 
   test 'reject multiple positions' do
-    result = ss.split('foo:bar:baz:quux', exclude: [1, -1])
+    result = ss.split('foo:bar:baz:quux', except: [1, -1])
     assert { result == %w[foo:bar baz:quux] }
 
     result = ss.split('foo:bar:baz:quux', reject: [1, -1])
     assert { result == %w[foo:bar baz:quux] }
 
-    result = ss.rsplit('foo:bar:baz:quux', exclude: [1, -1])
+    result = ss.rsplit('foo:bar:baz:quux', except: [1, -1])
     assert { result == %w[foo:bar baz:quux] }
 
     result = ss.rsplit('foo:bar:baz:quux', reject: [1, -1])